Move dict tools to their own module for better reuse.
